rag-mini-project/requirements.txt
# Add your Python dependencies here
fastapi
uvicorn
qdrant-client
sentence-transformers
streamlit
openai
python-dotenv
requests
nltk
python-multipart
chonkie
python-docx 
openpyxl 
PyMuPDF 
beautifulsoup4 
pandas 
python-pptx 
unstructured
python-docx

